requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
Flask-migrate
flask-cors
mysqlclient
flask-restful
python-dotenv
flask-jwt-extended
python-decouple
gunicorn